fixed import rules
